% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPSP.R
\name{SPSP}
\alias{SPSP}
\title{SPSP
Select the informative variables by the SPSP algorithm (Selection by Partitioning the Solution Paths)}
\usage{
SPSP(x, y, BETA, family = gaussian, init = 1, R = NULL, standardize = FALSE)
}
\arguments{
\item{x}{independent variables as a matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{response variable. Quantitative for \code{family="gaussian"} or \code{family="poisson"} (non-negative counts). 
For \code{family="binomial"} should be either a factor with two levels.}

\item{BETA}{a p by k matrix, should be sparser and sparser, each column corresponds to the estimated coefficients vector
for a given lambda, and lambda gets larger and larger.}

\item{family}{either a character string representing one of the built-in families, or else a glm() family object.}

\item{init}{initial coefficients, starting from init-th estimator}

\item{R}{sorted adjacent distances}

\item{standardize}{whether need standardization}
}
\value{

}
\description{
SPSP
Select the informative variables by the SPSP algorithm (Selection by Partitioning the Solution Paths)
}
\examples{
data(HihgDim)

}
