% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfun.SP.R
\name{Fitting-Functions}
\alias{Fitting-Functions}
\alias{lasso.glmnet}
\alias{lassoCV.glmnet}
\alias{adalasso.glmnet}
\alias{adalassoCV.glmnet}
\alias{adalassoCVmin.glmnet}
\alias{ridge.glmnet}
\alias{lasso.lars}
\alias{SCAD.ncvreg}
\alias{MCP.ncvreg}
\title{Nine Fitting-Functions that can be used as an input of \code{fitfun.SP} argument 
to obtain the solution paths for the SPSP algorithm.}
\usage{
lasso.glmnet(x, y, family, standardize, intercept, ...)

lassoCV.glmnet(x, y, family, standardize, intercept, ...)

adalasso.glmnet(x, y, family, standardize, intercept, ...)

adalassoCV.glmnet(x, y, family, standardize, intercept, ...)

adalassoCVmin.glmnet(x, y, family, standardize, intercept, ...)

ridge.glmnet(x, y, family, standardize, intercept, ...)

lasso.lars(x, y, family, standardize, intercept, ...)

SCAD.ncvreg(x, y, family, standardize, intercept, ...)

MCP.ncvreg(x, y, family, standardize, intercept, ...)
}
\arguments{
\item{x}{a matrix of the independent variables. The dimensions are (nobs) and (nvars); each row is an observation vector.}

\item{y}{Response variable. Quantitative for \code{family="gaussian"} or \code{family="poisson"} (non-negative counts). 
For \code{family="binomial"} should be either a factor with two levels.}

\item{family}{Response type. Either a character string representing one of the built-in families,
or else a glm() family object.}

\item{standardize}{logical argument. Should conduct standardization before the estimation? Default is TRUE.}

\item{intercept}{logical. If x is a data.frame, this argument determines if the resulting model matrix should contain 
a separate intercept or not. Default is TRUE.}

\item{...}{Additional optional arguments.}
}
\value{
An object of class \code{"glmnet"} is returned to provide solution paths for the SPSP algorithm.

An object of class \code{"glmnet"} is returned to provide solution paths for the SPSP algorithm.

An object of class \code{"glmnet"} is returned to provide solution paths for the SPSP algorithm.

An object of class \code{"glmnet"} using \code{lambda.1se} from the 10-fold 
cross-validation to provide solution paths for the SPSP algorithm.

An object of class \code{"glmnet"} using \code{lambda.min} from the 10-fold 
cross-validation to provide solution paths for the SPSP algorithm.

An object of class \code{"glmnet"} using ridge regression to provide solution
paths for the SPSP algorithm.

An object of class \code{"lars"} is returned to provide solution paths for the SPSP algorithm.

An object of class \code{"ncvreg"} to provide SCAD penalty solution paths for the SPSP algorithm.

An object of class \code{"ncvreg"} to provide MCP penalty solution paths for the SPSP algorithm.
}
\description{
The user can also customize a function to generate the solution paths, as long as 
the customized function take the arguments x, y, family, standardize, intercept, 
and return an object of class 
\code{glmnet}, \code{ncvreg}, and \code{lars}.
}
